Update `GradedTestSet` and `GradingResult` to use floats for points (#1)
* Update GradedTestSet and GradingResult to use Float64
instead of Int for points

* Fix category width calculation

---------

+++ src/GradedTestSets.jl
@@ -8,7 +8,7 @@ using Test: get_testset_depth, get_testset
 
 mutable struct GradedTestSet <: AbstractTestSet
     description::String
-    points::Int # points scored if all sub-tests pass
+    points::Float64 # points scored if all sub-tests pass
     results::Vector{Any}
 end
 GradedTestSet(desc; points=0) = GradedTestSet(desc, points, [])
